Primitive and Reference Variable
Write a program that demonstrates Primitive and Reference Variable by passing an integer and an array to a method, and modifying them in the method. Print the integer and the array in main , after they come out of the method. Observe how the integer is passed by value, but the array is passed by reference.