Task : Finding All Anagrams in a String
Objective: Given a string s and a non-empty string p, find all the start indices of p's anagrams in s.
Details: The goal is to develop a function that, given a string s and a non-empty string p, returns all the starting indices of the substrings in s that are anagrams of p. An anagram of a string is another string that contains the same characters, only the order of characters can be different. For example, "abc" is an anagram of "bca".
Functions to Implement:
List<Integer> findAnagrams(String s, String p): Returns a list of starting indices of the anagrams of p in s.